Field UI refactoring
